Consolidate duplicate functions
Functions & Support classes moved...
Robot_Kinematics, Interp, drawRobot2
