+++ scripts/generate_CenterPoints.py
@@ -24,6 +24,7 @@ from mpl_toolkits import mplot3d
 import random as random
 from mpl_toolkits.mplot3d import Axes3D
 
+from smart_golf_utilities import ROBOT_KINEMATICS, drawRobot2, Interp
 
 
 ######################
@@ -44,20 +45,9 @@ T = [np.pi,np.pi,np.pi,np.pi,np.pi,np.pi,np.pi]
 intervals = 16
 
 
-###############################
-## Support Class & Functions ##
-###############################
-
-def Interp(theta,intervals):
-    long_theta = [] 
-    parts = np.zeros((6,intervals-1))
-    for i in range(int(len(theta)-1)):
-        delta = (theta[i+1]-theta[i])/intervals
-        parts[i] = np.linspace(theta[i],theta[i+1]-delta,intervals-1)
-    for j in range(int(len(parts))):
-        long_theta = np.append(long_theta,parts[j])
-    
-    return long_theta
+##########
+## Main ##
+##########
 
 long_S = Interp(S,intervals)
 long_L = Interp(L,intervals)
@@ -65,121 +55,7 @@ long_U = Interp(U,intervals)
 long_R = Interp(R,intervals)
 long_B = Interp(B,intervals)
 long_T = Interp(T,intervals)
-                     
 
-class Robot():
-    
-     
-    """
-    This is the class for the wire frame robot used in all the scripts.
-    This is the way the forward kinematics are calculated for the hill climbs.
-    It is also useful for looking at the robot. 
-    """
-    #Class for the robot with lengths of links and the axis defined as they would be in the defalt zero position
-    
-    def __init__(self, links, axis = [[0,0,1],[0,1,0],[0,1,0],[0,1,0],[1,0,0]]):
-        self.links = links
-        self.axis = axis
-        
-    def rotateAxis(self, t, vec):
-        
-        #Rotation matrix just taken straight from wikepedia
-        
-        C = 1-np.cos(t)
-        x,y,z = vec
-
-        Rot = [[x**2*C+np.cos(t), x*y*C-z*np.sin(t), z*x*C+y*np.sin(t)],
-                [x*y*C+z*np.sin(t), y**2*C+np.cos(t), z*y*C-x*np.sin(t)],
-                [x*z*C-y*np. sin(t), y*z*C+x*np.sin(t), z**2*C+np.cos(t)]]
-
-        return Rot
-
-    def findEnd(self, angles):
-        #This function takes an input vector of angles and will output the end position as
-        #well as all of the vectors of the robot NOT TRANSLATED FROM THE ORIGIN
-        
-        v1 = [0,0,0] #Base thing that moves with S
-        v2 = [0,0,0] #First link that moves with L
-        v3 = [0,0,0] #Second link that moves with U
-        v4 = [0,0,0] #Shaft link that moves with B
-        v5 = [0,0,0] #Club head that moves with T/R
-        
-        #Starting Vector
-        v1_0 = [self.links[0]/np.sqrt(2),0,self.links[0]/np.sqrt(2)]
-        v5_0 = [0,self.links[4],0]
-        
-        #First rotation
-        v1 = np.dot(v1_0,self.rotateAxis(angles[0], self.axis[0]))
-        axis1 = np.dot(self.axis[1],self.rotateAxis(angles[0], self.axis[0]))
-        axis2 = np.dot(self.axis[2],self.rotateAxis(angles[0], self.axis[0]))
-        axis3 = np.dot(self.axis[3],self.rotateAxis(angles[0], self.axis[0]))
-        axis4 = np.dot(self.axis[4],self.rotateAxis(angles[0], self.axis[0]))
-        
-        #Second Rotation
-        v2 = (self.links[1]/self.links[0])*np.dot(v1,self.rotateAxis(angles[1], axis1))
-        axis2 = np.dot(axis2,self.rotateAxis(angles[1], axis1))
-        axis3 = np.dot(axis3,self.rotateAxis(angles[1], axis1))
-        axis4 = np.dot(axis4,self.rotateAxis(angles[1], axis1))
-        
-        #Third Rotation
-        v3 = (self.links[2]/self.links[1])*np.dot(v2,self.rotateAxis(angles[2], axis2))
-        axis3 = np.dot(axis3,self.rotateAxis(angles[2], axis2))
-        axis4 = np.dot(axis4,self.rotateAxis(angles[2], axis2))
-
-        #Fourth rotation
-        axis3 = np.dot(axis3,self.rotateAxis(angles[3], v3/self.links[2])) # Can only rotate around unit vectors
-        axis4 = np.dot(axis4,self.rotateAxis(angles[3], v3/self.links[2]))
-        
-        #Fifth Rotation
-        v4 = (self.links[3]/self.links[2])*np.dot(v3,self.rotateAxis(angles[4], axis3))
-        axis4 = np.dot(axis4,self.rotateAxis(angles[4], axis3))
-        
-        #Sixth Rotation
-        v5 = np.dot(v5_0,self.rotateAxis(angles[5], axis4))
-        
-        
-        x_val = v1[0]+v2[0]+v3[0]+v4[0]+v5[0]
-        y_val = v1[1]+v2[1]+v3[1]+v4[1]+v5[1]
-        z_val = v1[2]+v2[2]+v3[2]+v4[2]+v5[2]
-        
-        return [x_val, y_val, z_val, v1, v2, v3, v4, v5]
-    
-    def distanceFromTarget(self,targ,angles):
-        
-        #Simply finds the distance between a given end from a set of angles and a target end point
-        
-        end = self.findEnd(angles)
-        dist = (end[0]-targ[0])**2+(end[1]-targ[1])**2+(end[2]-targ[2])**2
-        
-        return dist
-    
-
-
-def drawRobot2(v1,v2,v3,v4,v5):
-    
-    """
-    The purpose of this function is just to convert the five vectors into
-    (x,y,z) arrays for plotting. 
-    
-    If you ever see drawRobot, that one is old so delete that
-    """
-    
-    x = np.zeros(6,)
-    y = np.zeros(6,)
-    z = np.zeros(6,)
-    
-    x[0],y[0],z[0] = [0,0,0]
-    x[1],y[1],z[1] = v1
-    x[2],y[2],z[2] = v1+v2
-    x[3],y[3],z[3] = v1+v2+v3
-    x[4],y[4],z[4] = v1+v2+v3+v4
-    x[5],y[5],z[5] = v1+v2+v3+v4+v5
-    
-    # print(x)
-    # print(y)
-    # print(z)
-    
-    return x,y,z
 
 # Make a real robot 
 real_links = np.array([88*np.sqrt(2),400,405,876,50.8]) #mm
@@ -195,7 +71,7 @@ anglesDesired = [S[3],L[3],U[3],R[3],B[3]-np.pi/4,T[3]]
 anglesConvention = [-1*anglesDesired[0],(np.pi/4-anglesDesired[1]),\
                     (-np.pi/2+anglesDesired[2]),anglesDesired[3],anglesDesired[4],np.pi-anglesDesired[5]]
 
-Moto = Robot(links = links_in, axis = axis)
+Moto = ROBOT_KINEMATICS(links = links_in, axis = axis)
 End = Moto.findEnd(anglesConvention)
 #print([End[0],End[1],End[2]])
